fixed layout and removed garbage of the project

+++ src/main/java/org/wanja/fatcat/map/Layer.java
@@ -10,5 +10,5 @@ public class Layer {
     public int x;
     public int y;
     public boolean visible;
-    public int[] data;
+    public long[] data;
 }
